Add notifications system, restructure components
