+++ src/Portal.tsx
@@ -21,9 +21,19 @@
  * @copyright SKALE Labs 2023-Present
  */
 
+import { useState, useEffect } from 'react'
+import { types } from '@/core'
+
 import Box from '@mui/material/Box'
 import CssBaseline from '@mui/material/CssBaseline'
-import { useMetaportStore, useWagmiAccount, Debug, cls, cmn } from '@skalenetwork/metaport'
+import {
+  useMetaportStore,
+  useWagmiAccount,
+  Debug,
+  cls,
+  cmn,
+  PROXY_ENDPOINTS
+} from '@skalenetwork/metaport'
 
 import Header from './Header'
 import SkDrawer from './SkDrawer'
@@ -31,17 +41,122 @@ import Router from './Router'
 import SkBottomNavigation from './SkBottomNavigation'
 import ProfileModal from './components/profile/ProfileModal'
 
+import { formatSChains } from './core/chain'
+import { STATS_API } from './core/constants'
+import { useSearchParams } from 'react-router-dom'
+import { getValidatorDelegations } from './core/delegation/staking'
+import { getValidator } from './core/delegation'
+import { initContracts } from './core/contracts'
+
 export default function Portal() {
   const mpc = useMetaportStore((state) => state.mpc)
+
+  const [schains, setSchains] = useState<types.ISChain[]>([])
+  const [metrics, setMetrics] = useState<types.IMetrics | null>(null)
+  const [stats, setStats] = useState<types.IStats | null>(null)
+  const [validator, setValidator] = useState<types.staking.IValidator | null | undefined>(null)
+  const [validatorDelegations, setValidatorDelegations] = useState<
+    types.staking.IDelegation[] | null
+  >(null)
+  const [customAddress, setCustomAddress] = useState<types.AddressType | undefined>(undefined)
+  const [sc, setSc] = useState<types.staking.ISkaleContractsMap | null>(null)
+  const [loadCalled, setLoadCalled] = useState<boolean>(false)
+
+  const endpoint = PROXY_ENDPOINTS[mpc.config.skaleNetwork]
+  const statsApi = STATS_API[mpc.config.skaleNetwork]
+
+  const [searchParams, _] = useSearchParams()
+
   const { address } = useWagmiAccount()
   if (!mpc) return <div></div>
+
+  useEffect(() => {
+    initSkaleContracts()
+    loadData()
+  }, [])
+
+  useEffect(() => {
+    loadValidator()
+  }, [address, customAddress, sc])
+
+  useEffect(() => {
+    setCustomAddress((searchParams.get('_customAddress') as types.AddressType) ?? undefined)
+  }, [location])
+
+  async function initSkaleContracts() {
+    setLoadCalled(true)
+    if (loadCalled) return
+    setSc(await initContracts(mpc))
+  }
+
+  async function loadChains() {
+    try {
+      const response = await fetch(`https://${endpoint}/files/chains.json`)
+      const chainsJson = await response.json()
+      setSchains(formatSChains(chainsJson))
+    } catch (e) {
+      console.log('Failed to load chains')
+      console.error(e)
+    }
+  }
+
+  async function loadMetrics() {
+    try {
+      const response = await fetch(`https://${endpoint}/files/metrics.json`)
+      const metricsJson = await response.json()
+      setMetrics(metricsJson)
+    } catch (e) {
+      console.log('Failed to load metrics')
+      console.error(e)
+    }
+  }
+
+  async function loadStats() {
+    if (statsApi === null) return
+    try {
+      const response = await fetch(statsApi)
+      const statsResp = await response.json()
+      setStats(statsResp.payload)
+    } catch (e) {
+      console.log('Failed to load stats')
+      console.error(e)
+    }
+  }
+
+  async function loadValidator() {
+    const addr = customAddress ?? address
+    if (!sc || !addr) return
+    const validatorData = await getValidator(sc.validatorService, addr)
+    setValidator(validatorData)
+    if (validatorData && validatorData.id) {
+      setValidatorDelegations(await getValidatorDelegations(sc, validatorData.id))
+    }
+  }
+
+  async function loadData() {
+    loadChains()
+    loadMetrics()
+    loadStats()
+    loadValidator()
+  }
+
   return (
     <Box sx={{ display: 'flex' }} className="AppWrap">
       <CssBaseline />
       <Header address={address} mpc={mpc} />
-      <SkDrawer />
+      <SkDrawer validatorDelegations={validatorDelegations} />
       <div className={cls(cmn.fullWidth)} id="appContentScroll">
-        <Router />
+        <Router
+          loadData={loadData}
+          schains={schains}
+          metrics={metrics}
+          stats={stats}
+          validator={validator}
+          validatorDelegations={validatorDelegations}
+          customAddress={customAddress}
+          sc={sc}
+          loadValidator={loadValidator}
+        />
         <ProfileModal />
         <div className={cls(cmn.mtop20, cmn.fullWidth)}>
           <Debug />
